switched segment compaction from compact-on-copy to compact-in-place

+++ cds/src/impl/krati/cds/impl/array/basic/ArrayEntryManager.java
@@ -11,34 +11,38 @@ import krati.cds.Persistable;
 import krati.cds.impl.array.entry.Entry;
 import krati.cds.impl.array.entry.EntryFactory;
 import krati.cds.impl.array.entry.EntryPersistListener;
+import krati.cds.impl.array.entry.EntryPool;
 import krati.cds.impl.array.entry.EntryValue;
+import krati.cds.impl.array.entry.PreFillEntryInt;
+import krati.cds.impl.array.entry.PreFillEntryLong;
+import krati.cds.impl.array.entry.PreFillEntryShort;
 
 public class ArrayEntryManager<V extends EntryValue> implements Persistable
 {
   private static final Logger _log = Logger.getLogger(ArrayEntryManager.class);
   
-  protected final int _maxEntries;
-  protected final int _maxEntrySize;
-  protected volatile boolean _autoSwitchEntry = true;    // Automatically switch to a new entry if _curEntry is full
-  protected volatile boolean _autoApplyEntries = true;   // Automatically apply accumulated entries to the array file
-  protected volatile long _lwmScn = 0;                   // Low water mark SCN starts from 0
-  protected volatile long _hwmScn = 0;                   // High water mark SCN starts from 0
+  private final int _maxEntries;
+  private final int _maxEntrySize;
+  private volatile boolean _autoApplyEntries = true;   // Automatically apply accumulated entries to the array file
+  private volatile long _lwmScn = 0;                   // Low water mark SCN starts from 0
+  private volatile long _hwmScn = 0;                   // High water mark SCN starts from 0
   
-  protected RecoverableArray<V>  _array;
-  protected List<Entry<V>>       _entryList;      // list of redo entry(s)
-  protected List<Entry<V>>       _clonedEntryList;// list of cloned redo entry(s)
-  protected Entry<V>             _currentEntry;   // current redo entry
-  protected EntryPersistListener _persistListener;
+  private RecoverableArray<V>  _array;
+  private Entry<V>             _entry;                 // current redo entry
+  private final EntryPool<V>   _entryPool;
+  private final EntryApply<V>  _entryApply;
+  private EntryPersistListener _persistListener;
   
   public ArrayEntryManager(RecoverableArray<V> array, int maxEntries, int maxEntrySize)
   {
     this._array = array;
     this._maxEntries = maxEntries;
     this._maxEntrySize = maxEntrySize;
-    this._entryList = new ArrayList<Entry<V>>(maxEntries + 5);
-    this._clonedEntryList = new ArrayList<Entry<V>>(maxEntries + 5);
+    this._entryPool = new EntryPool<V>(array.getEntryFactory(), maxEntrySize);
+    this._entryApply = new EntryApply<V>(this);
+    this._entry = _entryPool.next();
     
-    _log.info("Manage array indexStart=" + array.getIndexStart() + " length=" + array.length());
+    _log.info("arrayLength=" + array.length() + " maxEntries=" + maxEntries + " maxEntrySize=" + maxEntrySize);
   }
   
   public int getMaxEntries()
@@ -61,16 +65,6 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
     return _array.getEntryFactory();
   }
   
-  public boolean getAutoSwitchEntry()
-  {
-    return _autoSwitchEntry;
-  }
-  
-  public void setAutoSwitchEntry(boolean  b)
-  {
-    _autoSwitchEntry = b;
-  }
-  
   public boolean getAutoApplyEntries()
   {
     return _autoApplyEntries;
@@ -81,32 +75,72 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
     _autoApplyEntries = b;
   }
   
-  public void addToEntry(V entryValue) throws IOException
+  void addToEntry(V entryValue) throws IOException
   {
-    if (_currentEntry == null)
-    {
-      _currentEntry = createEntry();
-    }
-    
     // Add to current entry
-    _currentEntry.add(entryValue);
+    _entry.add(entryValue);
     
     // Advance high water mark to maintain progress record
     _hwmScn = Math.max(_hwmScn, entryValue.scn);
     
     // Switch to a new entry if _currentEntry has reached _maxEntrySize and has seen more than one SCN.
-    if (_autoSwitchEntry && _currentEntry.isSwitchable())
+    if(_entry.isFull())
     {
       switchEntry(false);
     }
+  }
+  
+  void addToPreFillEntryInt(int pos, int val, long scn) throws IOException
+  {
+      // Add to current entry
+      ((PreFillEntryInt)_entry).add(pos, val, scn);
+      
+      // Advance high water mark to maintain progress record
+      _hwmScn = Math.max(_hwmScn, scn);
+      
+      // Switch to a new entry if _currentEntry has reached _maxEntrySize and has seen more than one SCN.
+      if(_entry.isFull())
+      {
+        switchEntry(false);
+      }
+  }
+  
+  void addToPreFillEntryLong(int pos, long val, long scn) throws IOException
+  {
+      // Add to current entry
+      ((PreFillEntryLong)_entry).add(pos, val, scn);
+      
+      // Advance high water mark to maintain progress record
+      _hwmScn = Math.max(_hwmScn, scn);
+      
+      // Switch to a new entry if _currentEntry has reached _maxEntrySize and has seen more than one SCN.
+      if(_entry.isFull())
+      {
+        switchEntry(false);
+      }
+  }
+  
+  void addToPreFillEntryShort(int pos, short val, long scn) throws IOException
+  {
+   // Add to current entry
+      ((PreFillEntryShort)_entry).add(pos, val, scn);
+      
+      // Advance high water mark to maintain progress record
+      _hwmScn = Math.max(_hwmScn, scn);
+      
+      // Switch to a new entry if _currentEntry has reached _maxEntrySize and has seen more than one SCN.
+      if(_entry.isFull())
+      {
+        switchEntry(false);
+      }
   }
   
   public synchronized void clear()
   {
     _lwmScn = 0;
     _hwmScn = 0;
-    _entryList.clear();
-    _currentEntry = null;
+    _entry.clear();
+    _entryPool.clear();
     
     try
     {
@@ -176,27 +210,12 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
     return _persistListener;
   }
   
-  public File getEntryLogFile(String tag)
-  {
-    // history_<id>_<tag>.idx
-    return new File(getCacheDirectory(), getEntryLogPrefix() + "_" + tag + getEntryLogSuffix());
-  }
-  
-  /**
-   * Create a new entry.
-   * @return a new entry.
-   */
-  protected Entry<V> createEntry()
-  {
-    return getEntryFactory().newEntry(_maxEntrySize);
-  }
-  
   /**
    * @return the name of entry log file.
    */
   protected final String getEntryLogName(Entry<V> entry)
   {
-    return getEntryLogPrefix() + "_" + entry.getMinScn() + "_" + entry.getMaxScn() + getEntryLogSuffix();
+    return getEntryLogPrefix() + "_" + entry.getServiceId() + "_" + entry.getMinScn() + "_" + entry.getMaxScn() + getEntryLogSuffix();
   }
   
   /**
@@ -204,7 +223,7 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
    */
   protected final String getEntryLogPrefix()
   {
-    return "history_" + _array.getIndexStart();
+    return "entry";
   }
   
   /**
@@ -216,56 +235,44 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
   }
   
   /**
-   * @return a list of accumulated entries.
-   */
-  protected List<Entry<V>> getEntryList()
-  {
-    return _entryList;
-  }
-  
-  /**
    * Switches to a new entry if _curEntry is not empty.
    * 
    * @throws IOException
    */
   protected synchronized void switchEntry(boolean blocking) throws IOException
   {
-    _log.info("switchEntry ...");
-    
-    if (_currentEntry != null && !_currentEntry.isEmpty())
+    if (!_entry.isEmpty())
     {
       if(_persistListener != null)
       {
-          _persistListener.priorPersisting(_currentEntry);
+        _persistListener.priorPersisting(_entry);
       }
       
       // Create entry log and persist in-memory data
-      File file = new File(getCacheDirectory(), getEntryLogName(_currentEntry));
-      _currentEntry.save(file);
+      File file = new File(getCacheDirectory(), getEntryLogName(_entry));
+      _entry.save(file);
       
       if(_persistListener != null)
       {
-          _persistListener.afterPersisting(_currentEntry);
+        _persistListener.afterPersisting(_entry);
       }
       
       // Advance low water mark to maintain progress record
-      _lwmScn = Math.max(_lwmScn, _currentEntry.getMaxScn());
-      _entryList.add(_currentEntry);
-      _currentEntry = null;
+      _lwmScn = Math.max(_lwmScn, _entry.getMaxScn());
+      _entryPool.addToServiceQueue(_entry);
+      _entry = _entryPool.next();
       
-      _log.info("_lwmScn=" + _lwmScn + " _hwmScn=" + _hwmScn);
+      _log.info("switchEntry to " + _entry.getId() + " _lwmScn=" + _lwmScn + " _hwmScn=" + _hwmScn);
     }
     
     // Apply entry logs to array file
     if (_autoApplyEntries)
     {
-      if (_entryList.size() >= _maxEntries)
+      if (_entryPool.getServiceQueueSize() >= _maxEntries)
       {
         applyEntries(blocking);
       }
     }
-    
-    _log.info("switchEntry end");
   }
   
   /**
@@ -274,69 +281,58 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
    */
   protected synchronized void applyEntries(boolean blocking) throws IOException
   {
-    if (_entryList.size() > 0)
+    if (blocking) /* Blocking Mode */
     {
-      synchronized(_clonedEntryList)
+      synchronized(_entryApply)
       {
-        while(_clonedEntryList.size() > 0)
+        List<Entry<V>> entryList = new ArrayList<Entry<V>>();
+        while(_entryPool.getServiceQueueSize() > 0)
         {
-          try
-          { 
-            _clonedEntryList.wait();
-          }
-          catch(InterruptedException ie)
-          {
-            /* Run in blocking mode */
-            
-            // Update underlying array file
-            _array.updateArrayFile(_entryList);
-              
-            // Clean up entry files
-            deleteEntryFiles(_entryList);
-            _entryList.clear();
-            return;
-          }
+          Entry<V> entry = _entryPool.pollFromService();
+          if(entry != null) entryList.add(entry);
         }
         
-        // Blocking mode
-        if (blocking)
+        applyEntries(entryList);
+      }
+    }
+    else      /* Non-Blocking Mode */
+    {
+      synchronized(_entryApply)
+      {
+        for(int i = 0; i < _maxEntries; i++)
         {
-          // Update underlying array file
-          _array.updateArrayFile(_entryList);
-          
-          // Clean up entry files
-          deleteEntryFiles(_entryList);
-          _entryList.clear();
-        }
-        // Non-blocking mode
-        else
-        {
-          // Start a separate thread to update the underlying array file
-          for(Entry<V> e: _entryList) { _clonedEntryList.add(e); }
-          new Thread(new UpdateRunner()).run();
-          _entryList.clear();
+          Entry<V> entry = _entryPool.pollFromService();
+          if(entry != null) _entryApply.add(entry);
         }
       }
+      
+      // Start a separate thread to update the underlying array file 
+      new Thread(_entryApply).run();
     }
   }
   
-  class UpdateRunner implements Runnable
-  { 
+  static class EntryApply<V extends EntryValue> implements Runnable
+  {
+    private final List<Entry<V>> _entryList;
+    private final ArrayEntryManager<V> _entryManager;
+    
+    public EntryApply(ArrayEntryManager<V> entryManager)
+    {
+      _entryManager = entryManager;
+      _entryList = new ArrayList<Entry<V>>();
+    }
+    
+    public final void add(Entry<V> entry)
+    {
+      _entryList.add(entry);
+    }
+    
     @Override
-    public void run()
+    public final synchronized void run()
     {
       try
       {
-        synchronized(_clonedEntryList)
-        {
-          // Update underlying array file
-          _array.updateArrayFile(_clonedEntryList);
-          
-          // Clean up entry files
-          deleteEntryFiles(_clonedEntryList);
-          _clonedEntryList.clear();  
-          _clonedEntryList.notifyAll();
-        }
+        _entryManager.applyEntries(_entryList);
       }
       catch(IOException ioe)
       {
@@ -350,11 +346,13 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
    * 
    * @throws IOException
    */
-  protected void loadEntryFiles()
+  protected List<Entry<V>> loadEntryFiles()
   {
     File[] files = getCacheDirectory().listFiles();
     String prefix = getEntryLogPrefix();
     String suffix = getEntryLogSuffix();
+    
+    List<Entry<V>> entryList = new ArrayList<Entry<V>>();
     
     for (File file : files)
     {
@@ -363,9 +361,9 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
       {
         try
         {
-          Entry<V> entry = createEntry();
+          Entry<V> entry = _entryPool.next();
           entry.load(file);
-          _entryList.add(entry);
+          entryList.add(entry);
         }
         catch(IOException e)
         {
@@ -378,6 +376,8 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
         }
       }
     }
+    
+    return entryList;
   }
   
   /**
@@ -413,37 +413,37 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
    * 
    * @throws IOException
    */
-  protected void deleteEntryFiles(List<Entry<V>> list) throws IOException
+  private void deleteEntryFiles(List<Entry<V>> list) throws IOException
   {
     for(Entry<V> e: list)
     {
-      File file = new File(getCacheDirectory(), getEntryLogName(e));
-      if(file.exists())
+      File file = e.getFile();
+      if(file != null && file.exists())
       {
         if (file.delete())
         {
-          _log.info("file " + file.getAbsolutePath() + " deleted");
+          _log.info(file.getName() + " deleted");
         }
         else 
         {
-          _log.warn("file " + file.getAbsolutePath() + " not deleted");
+          _log.warn(file.getName() + " not deleted");
         }
       }
     }
   }
   
   /**
-   * Filter and select entries in _entryList
+   * Filter and select entries from an entry list
    * that only have SCNs no less than the specified lower bound minScn and
    * that only have SCNs no greater than the specified upper bound maxScn.
    * 
    * @param minScn Inclusive lower bound SCN.
    * @param maxScn Inclusive upper bound SCN.
    */
-  protected void filterEntryList(long minScn, long maxScn)
+  private List<Entry<V>> filterEntryList(List<Entry<V>> entryList, long minScn, long maxScn)
   {
-    List<Entry<V>> result = new ArrayList<Entry<V>>(_entryList.size());
-    for (Entry<V> e : _entryList)
+    List<Entry<V>> result = new ArrayList<Entry<V>>(entryList.size());
+    for (Entry<V> e : entryList)
     {
       if (minScn <= e.getMinScn() && e.getMaxScn() <= maxScn)
       {
@@ -451,22 +451,18 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
       }
     }
     
-    if (result.size() < _entryList.size())
-    {
-      _entryList.clear();
-      _entryList.addAll(result);
-    }
+    return result;
   }
   
   /**
-   * Filter and select entries in _entryList that only have SCNs no less than the specified lower bound.
+   * Filter and select entries from an entry list that only have SCNs no less than the specified lower bound.
    * 
    * @param scn Inclusive lower bound SCN
    */
-  protected void filterEntryListLowerBound(long scn)
+  private List<Entry<V>> filterEntryListLowerBound(List<Entry<V>> entryList, long scn)
   {
-    List<Entry<V>> result = new ArrayList<Entry<V>>(_entryList.size());
-    for (Entry<V> e : _entryList)
+    List<Entry<V>> result = new ArrayList<Entry<V>>(entryList.size());
+    for (Entry<V> e : entryList)
     {
       if (scn <= e.getMinScn())
       {
@@ -474,22 +470,19 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
       }
     }
     
-    if (result.size() < _entryList.size())
-    {
-      _entryList.clear();
-      _entryList.addAll(result);
-    }
+    return result;
   }
 
   /**
-   * Filter and select entries in _entryList that only have SCNs no greater than the specified upper bound.
+   * Filter and select entries from an entry list that only have SCNs no greater than the specified upper bound.
    * 
    * @param scn Inclusive upper bound SCN
    */
-  protected void filterEntryListUpperBound(long scn)
+  @SuppressWarnings("unused")
+  private List<Entry<V>> filterEntryListUpperBound(List<Entry<V>> entryList, long scn)
   {
-    List<Entry<V>> result = new ArrayList<Entry<V>>(_entryList.size());
-    for (Entry<V> e : _entryList)
+    List<Entry<V>> result = new ArrayList<Entry<V>>(entryList.size());
+    for (Entry<V> e : entryList)
     {
       if (e.getMaxScn() <= scn)
       {
@@ -497,11 +490,61 @@ public class ArrayEntryManager<V extends EntryValue> implements Persistable
       }
     }
     
-    if (result.size() < _entryList.size())
-    {
-      _entryList.clear();
-      _entryList.addAll(result);
-    }
+    return result;
   }
-
+  
+  protected void applyEntries(List<Entry<V>> entries) throws IOException
+  { 
+    // Update underlying array file
+    _array.updateArrayFile(entries);
+    
+    // Clean up entry files
+    deleteEntryFiles(entries);
+    
+    // Recycle all applied entries
+    for(Entry<V> entry : entries)
+    {
+      _entryPool.addToRecycleQueue(entry);
+    }
+    
+    entries.clear();
+  }
+  
+  protected void init(long arrayFileMaxScn, long arrayFileNewScn) throws IOException
+  {
+    // Load entries from logs on disk 
+    List<Entry<V>> entryList = loadEntryFiles();
+    
+    // Sanitize loaded entries
+    if (arrayFileMaxScn == arrayFileNewScn) // array file is consistent 
+    {
+      // Find entries that have not been flushed to the array file.
+      if(entryList.size() > 0)
+      {
+        entryList = filterEntryListLowerBound(entryList, arrayFileMaxScn);
+      }
+    }
+    else // array file is inconsistent
+    {
+      if(entryList.size() > 0)
+      {
+        entryList = filterEntryList(entryList, arrayFileMaxScn, arrayFileNewScn);
+        
+        if(entryList.size() == 0)
+        {
+          deleteEntryFiles();
+          _log.error("entry files for recovery not found");
+        }
+      }
+    }
+    
+    // Start recovery based on loaded entries.
+    if (entryList.size() > 0)
+    {
+      applyEntries(entryList);
+    }
+    
+    // Delete whatever entry files on disk.
+    deleteEntryFiles();
+  }
 }
